refactor(core): 🔨 split dns tunneling and ntp helpers
