+++ DomainDetective/DomainHealthCheck.Verification.cs
@@ -330,33 +330,6 @@ namespace DomainDetective {
             await IPNeighborAnalysis.Analyze(domainName, _logger, cancellationToken);
         }
 
-        /// <summary>Analyzes DNS logs for tunneling patterns.</summary>
-        public void CheckDnsTunneling(string domainName, CancellationToken ct = default) {
-            CheckDnsTunnelingAsync(domainName, ct).GetAwaiter().GetResult();
-        }
-
-        public async Task CheckDnsTunnelingAsync(string domainName, CancellationToken ct = default) {
-            ct.ThrowIfCancellationRequested();
-            var lines = DnsTunnelingLogs ?? Array.Empty<string>();
-            await Task.Run(() => DnsTunnelingAnalysis.Analyze(domainName, lines), ct);
-        }
-
-        /// <summary>Queries an NTP server for clock information.</summary>
-        /// <param name="host">Target server host name or IP.</param>
-        /// <param name="port">NTP port number.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public async Task TestNtpServer(string host, int port = 123, CancellationToken cancellationToken = default) {
-            ValidatePort(port);
-            await NtpAnalysis.AnalyzeServer(host, port, _logger, cancellationToken);
-        }
-
-        /// <summary>Queries a predefined NTP server.</summary>
-        /// <param name="server">Built-in server enumeration.</param>
-        /// <param name="port">NTP port number.</param>
-        /// <param name="cancellationToken">Token to cancel the operation.</param>
-        public Task TestNtpServer(NtpServer server, int port = 123, CancellationToken cancellationToken = default) =>
-            TestNtpServer(server.ToHost(), port, cancellationToken);
-
